refactor: extract shared test utilities into git::test_utils
branch.rs、diff.rs、worktree.rs に重複していたテスト初期化関数を
src/git/test_utils.rs に集約。各テストモジュールから共通ヘルパーを
使うよう修正。

Closes #88
