+++ src/git/mod.rs
@@ -1,5 +1,7 @@
 pub mod branch;
 pub mod diff;
+#[cfg(test)]
+pub mod test_utils;
 pub mod worktree;
 
 use anyhow::{Context, Result};
